:recycle: Split study schema into queries
